Add start and stop analytics functions
